PUB-2491 - Added tests

+++ src/main/controllers/RemoveListConfirmationController.ts
@@ -7,6 +7,7 @@ import { UserManagementService } from '../service/UserManagementService';
 import { addListDetailsToArray } from '../helpers/listHelper';
 import { RemoveListHelperService } from '../service/RemoveListHelperService';
 import * as url from 'url';
+import {checkIfUrl} from "../helpers/urlHelper";
 
 const courtService = new LocationService();
 const manualUploadService = new ManualUploadService();
@@ -39,7 +40,7 @@ export default class RemoveListConfirmationController {
         const locationId = req.body?.locationId;
         const formData = req.body;
         const listData = [];
-        if (listsToDelete && locationId) {
+        if (listsToDelete && locationId && !checkIfUrl(locationId)) {
             switch (formData['remove-choice']) {
                 case 'yes': {
                     const response = await removeListHelperService.removeLists(listsToDelete, req.user?.['userId']);
